+++ app/javascript/components/SignUp/DefaultForm/DefaultForm.scss
@@ -1,0 +1,12 @@
+.button {
+    height: 60px;
+    margin-top: 20px;
+    margin-bottom: 20px;
+    width: 100%;
+}
+
+.agreement {
+    font-weight: 400;
+    font-size: 12px;
+    color: grey;
+}
