+++ app/javascript/components/SignUp/DefaultForm/index.js
@@ -1,0 +1,1 @@
+export { DefaultForm } from './DefaultForm';
