+++ app/javascript/components/SignUp/DriverForm/index.js
@@ -1,0 +1,1 @@
+export { DriverForm } from './DriverForm';
