+++ app/javascript/components/SignUp/SignUp.scss
@@ -20,20 +20,11 @@
 }
 
 .title-two {
-    margin-bottom: 40px;
-}
-
-.agreement {
-    font-weight: 400;
-    font-size: 12px;
-    color: grey;
-}
-
-.button {
-    height: 60px;
-    margin-top: 20px;
     margin-bottom: 20px;
-    width: 100%;
+}
+
+.driver {
+    margin-bottom: 20px;
 }
 
 @media (max-width: 568px) {
